+++ src/file_list.cmake
@@ -31,6 +31,7 @@ set (file_list
     ${CMAKE_CURRENT_LIST_DIR}/customprops/font_string_prop.cpp    # Derived wxStringProperty class for font property
     ${CMAKE_CURRENT_LIST_DIR}/customprops/html_string_prop.cpp    # Derived wxStringProperty class for HTML
     ${CMAKE_CURRENT_LIST_DIR}/customprops/img_string_prop.cpp     # Derived wxStringProperty class for handling wxImage files or art
+    ${CMAKE_CURRENT_LIST_DIR}/customprops/rearrange_prop.cpp      # wxRearangeList contents editor
     ${CMAKE_CURRENT_LIST_DIR}/customprops/sb_fields_prop.cpp      # Property editor for status bar fields
     ${CMAKE_CURRENT_LIST_DIR}/customprops/txt_string_prop.cpp     # Derived wxStringProperty class for single-line text